RawAPI and clear view! 🦅 Seperated `RawAPI` from `Televerse`
